rss-producer/requirements.txt
kafka-python
feedparser

